# server/api-lib/docs/openapi.admin.yaml
openapi: 3.1.0
info:
  title: Explore Admin HTTP API
  version: 1.0.0
  description: |
    Hand-authored OpenAPI for the Explore web `/api/*` surface (Vercel `router.mjs`
    + local Express). PostgREST and Edge Functions are separate Scalar sources.

    `POST /api/events` matches `docs/ANALYTICS_EVENTS_API.md`.
servers:
  - url: /
    description: Same origin (Vercel production or local Vite proxy)
...
security: []
paths:
...
  /api/reports:
    post:
      tags: [Moderation]
      summary: Submit a content report
      operationId: postReports
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Report accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/user/hidden-content:
    get:
      tags: [Moderation]
      summary: List content hidden by the current user
      operationId: getUserHiddenContent
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Hidden content list
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Moderation]
      summary: Hide content for the current user
      operationId: postUserHiddenContent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Content hidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/user/hidden-content/unhide:
    post:
      tags: [Moderation]
      summary: Unhide previously hidden content
      operationId: postUserHiddenContentUnhide
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Content unhidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/admin/me:
    get:
      tags: [Moderation]
      summary: Current admin identity
      operationId: getAdminMe
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Admin profile
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/admins:
    get:
      tags: [Moderation]
      summary: List admin roster
      operationId: getAdminRoster
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Admin users from admin_users
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Moderation]
      summary: Add or upsert an admin by email
      operationId: postAdminRoster
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                role:
                  type: string
                  enum: [admin, moderator]
      responses:
        "200":
          description: Admin upserted
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Auth user not found for email
    patch:
      tags: [Moderation]
      summary: Update an admin role
      operationId: patchAdminRoster
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, role]
              properties:
                user_id:
                  type: string
                  format: uuid
                role:
                  type: string
                  enum: [admin, moderator]
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [Moderation]
      summary: Remove an admin from the roster
      operationId: deleteAdminRoster
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id:
                  type: string
                  format: uuid
      responses:
        "200":
          description: Admin removed
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/reports:
    get:
      tags: [Moderation]
      summary: List moderation reports
      operationId: getAdminReports
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Report queue
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/reports/{id}:
    get:
      tags: [Moderation]
      summary: Get a moderation report by id
      operationId: getAdminReportById
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ReportId"
      responses:
        "200":
          description: Report detail
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Moderation]
      summary: Update a moderation report
      operationId: patchAdminReportById
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ReportId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Report updated
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/moderation/summary:
    get:
      tags: [Moderation]
      summary: Moderation summary counters
      operationId: getAdminModerationSummary
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Summary payload
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/ops/summary:
    get:
      tags: [Moderation]
      summary: Ops summary for the admin console
      operationId: getAdminOpsSummary
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Ops summary
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/users:
    get:
      tags: [Moderation]
      summary: List admin-visible users
      operationId: getAdminUsers
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/moderation/action:
    post:
      tags: [Moderation]
      summary: Apply a moderation action
      operationId: postAdminModerationAction
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Action applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/events:
    post:
      tags: [Analytics]
      summary: Ingest analytics event batch
      description: |
        Mobile/web clients send analytics batches. Contract mirrors
        `docs/ANALYTICS_EVENTS_API.md`. Optional bearer token; anonymous
        batches require `anonymous_id`. Max 50 events / 256 KB.
      operationId: postAnalyticsEvents
      security:
        - bearerAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyticsEventsBatch"
            example:
              batch_id: batch-uuid
              sent_at: "2026-07-02T10:00:00.000Z"
              events:
                - event_id: event-uuid
                  event_name: video_view_start
                  event_version: 1
                  anonymous_id: stable-anonymous-id
                  session_id: session-id
                  entity_type: video
                  entity_id: video-id
                  occurred_at: "2026-07-02T09:59:58.000Z"
                  platform: ios
                  app_version: "1.0.0"
                  build_number: "100"
                  locale: es-DO
                  timezone: America/Santo_Domingo
                  country: DO
                  properties: {}
                  context: {}
      responses:
        "200":
          description: Batch processed (may include partial rejects)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsEventsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          description: Payload too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Analytics schema or Supabase not ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/cron/analytics/aggregate:
    post:
      tags: [Analytics]
      summary: Cron trigger for analytics aggregates
      operationId: postCronAnalyticsAggregate
      security:
        - cronSecretAuth: []
      responses:
        "200":
          description: Aggregation completed
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/admin/analytics/events/{eventId}:
    get:
      tags: [Admin Analytics]
      summary: Get a single analytics event
      operationId: getAdminAnalyticsEventById
      security:
        - bearerAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Event detail
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/admin/system/bootstrap-board:
    post:
      tags: [System]
      summary: Bootstrap board admin accounts
      operationId: postBootstrapBoardAdmins
      security:
        - bearerAuth: []
        - bootstrapSecretAuth: []
      responses:
        "200":
          description: Bootstrap result
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/waitlist/signup:
    post:
      tags: [Waitlist]
      summary: Public waitlist signup
      operationId: postWaitlistSignup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                phone:
                  type: string
              additionalProperties: false
      responses:
        "200":
          description: Signup result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "400":
          description: Missing email/phone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/admin/waitlist:
    get:
      tags: [Waitlist]
      summary: List waitlist entries
      operationId: getAdminWaitlist
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Waitlist rows + stats
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/waitlist/notify-launch:
    post:
      tags: [Waitlist]
      summary: Email waitlist that the app is ready
      operationId: postAdminWaitlistNotifyLaunch
      security:
        - bearerAuth: []
      parameters:
        - name: dryRun
          in: query
          schema:
            type: string
            enum: ["1"]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                dryRun:
                  type: boolean
              additionalProperties: true
      responses:
        "200":
          description: Send result
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/broadcast:
    post:
      tags: [Waitlist]
      summary: Broadcast email/SMS to waitlist
      operationId: postAdminBroadcast
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Broadcast result
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase user access token (admin routes require admin_users).
    metricsTokenAuth:
      type: http
      scheme: bearer
      description: METRICS_TOKEN bearer for Prometheus scrapes.
    cronSecretAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Bearer ANALYTICS_CRON_SECRET (or project cron auth).
    bootstrapSecretAuth:
      type: apiKey
      in: header
      name: x-admin-bootstrap-secret
      description: Optional ADMIN_BOOTSTRAP_SECRET for board bootstrap.
  parameters:
    ReportId:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
        request_id:
          type: string
      additionalProperties: true
    ErrorResponse:
      type: object
      properties:
        ok:
          type: boolean
          const: false
        error:
          type: string
        code:
          type: string
        request_id:
          type: string
      additionalProperties: true
    AnalyticsEventsBatch:
      type: object
      required: [events]
      properties:
        batch_id:
          type: string
        sent_at:
          type: string
          format: date-time
        events:
          type: array
          maxItems: 50
          items:
            $ref: "#/components/schemas/AnalyticsEvent"
      additionalProperties: false
...