# server/index.ts
import cors from "cors";
import express from "express";
import type { Request as ExpressRequest, Response as ExpressResponse } from "express";
import jwt from "jsonwebtoken";
// Plain ESM handlers are shared with Vercel serverless functions.
import {
  dispatchAdminAnalyticsApi,
  handleCronAnalyticsAggregate,
} from "./api-lib/analytics/analyticsAdminApi.mjs";
// @ts-ignore
import { dispatchBusinessIntelligenceApi } from "./api-lib/analytics/businessIntelligenceApi.mjs";
// @ts-ignore
import { handleEvents } from "./api-lib/analytics/analyticsRouter.mjs";
// @ts-ignore
import {
  handleAdminModerationAction,
  handleAdminModerationSummary,
  handleAdminMe,
  handleAdminOpsSummary,
  handleAdminReportById,
  handleAdminReports,
  handleAdminUsers,
  handleHealth,
  handleReports,
  handleUserHiddenContent,
  handleUserHiddenContentUnhide,
} from "./api-lib/moderation/supabaseModeration.mjs";
// @ts-ignore
import { handleAdminRoster } from "./api-lib/moderation/adminRosterApi.mjs";
// @ts-ignore
import {
  handleAdminSystemHealth,
  handleAdminSystemMetrics,
  handleBootstrapBoardAdmins,
  handleTokenMetrics,
} from "./api-lib/system/systemRouter.mjs";
// @ts-ignore
import { handlePioneersLanding } from "./api-lib/pioneers/pioneersRouter.mjs";
// @ts-ignore
import { dispatchOpenApiDocs } from "./api-lib/docs/docsRouter.mjs";
// @ts-ignore
import adminBroadcast from "../netlify/functions/admin-broadcast.mjs";
// @ts-ignore
import feedbackSubmit from "../netlify/functions/feedback-submit.mjs";
// @ts-ignore
import { ensureRequestId, routePath } from "./api-lib/http/requestContext.mjs";
// @ts-ignore
import { resolveApiRoute } from "./api-lib/http/resolveApiRoute.mjs";
// @ts-ignore
import { errorSummary, logger } from "./api-lib/observability/logger.mjs";
// @ts-ignore
import { recordApiRequest } from "./api-lib/observability/metrics.mjs";
// @ts-ignore
import { observabilityConfigStatus, flushPendingLokiLogs, runWithObservabilityContext } from "./api-lib/observability/lokiLogger.mjs";
import { requireAdmin } from "./adminAuth.js";
import { config } from "./config.js";
import {
  addToWaitlist,
  getWaitlistStats,
  listWaitlist,
  listWaitlistPendingLaunch,
  markLaunchNotified,
} from "./db.js";
import { isFirebaseAdminConfigured } from "./firebaseAdmin.js";
import { sendAppLaunchBulk, sendWaitlistEmail } from "./mail.js";
import {
  listPendingLaunchEmailsMergedAsync,
  listWaitlistMergedAsync,
} from "./waitlistMerged.js";
import { markLaunchNotifiedFirestore } from "./waitlistFirestore.js";

const app = express();
app.use(cors({ origin: true, credentials: true }));
app.use(express.json());

// Request-scoped Loki flush for local Express (same enqueue path as Vercel waitUntil).
app.use((_req, res, next) => {
  runWithObservabilityContext(() => {
    res.on("finish", () => {
      void flushPendingLokiLogs({ timeoutMs: 1500 });
    });
    next();
  });
});

type FetchHandler = (request: Request) => Promise<Response>;

function expressHeaders(req: ExpressRequest) {
  const headers = new Headers();
  for (const [key, value] of Object.entries(req.headers)) {
    if (value == null) continue;
    if (Array.isArray(value)) {
      value.forEach((item) => headers.append(key, item));
    } else {
      headers.set(key, value);
    }
  }
  return headers;
}

function expressUrl(req: ExpressRequest) {
  const protocol = req.headers["x-forwarded-proto"] || req.protocol || "http";
  const host = req.headers["x-forwarded-host"] || req.headers.host || `localhost:${config.port}`;
  return `${protocol}://${host}${req.originalUrl || req.url}`;
}

async function sendFetchResponse(
  handler: FetchHandler,
  req: ExpressRequest,
  res: ExpressResponse,
) {
  const started = Date.now();
  let routeLabel = "/api";
  try {
    const headers = expressHeaders(req);
    let body: string | undefined;
    if (req.method !== "GET" && req.method !== "HEAD" && req.body !== undefined) {
      body = typeof req.body === "string" ? req.body : JSON.stringify(req.body);
      if (!headers.has("content-type")) headers.set("content-type", "application/json");
    }

    const request = new Request(expressUrl(req), {
      method: req.method,
      headers,
      body,
    });
    const { request: requestWithId } = ensureRequestId(request);
    routeLabel = routePath(resolveApiRoute(requestWithId));
    const response = await handler(requestWithId);

    res.status(response.status);
    response.headers.forEach((value, key) => {
      if (key.toLowerCase() !== "transfer-encoding") {
        res.setHeader(key, value);
      }
    });

    const buffer = Buffer.from(await response.arrayBuffer());
    if (buffer.length > 0) res.send(buffer);
    else res.end();

    recordApiRequest({
      route: routeLabel,
      method: req.method,
      status: response.status,
      durationMs: Date.now() - started,
    });
  } catch (err) {
    logger.error("Express API handler failed", {
      route: req.originalUrl || req.url,
      method: req.method,
      error: errorSummary(err),
    });
    recordApiRequest({
      route: routeLabel,
      method: req.method,
      status: 500,
      durationMs: Date.now() - started,
    });
    res.status(500).json({ ok: false, error: "Unexpected server error." });
  }
}

function normalizeEmail(email: unknown): string | null {
  if (typeof email !== "string") return null;
  const trimmed = email.trim().toLowerCase();
  if (!/^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(trimmed)) return null;
  return trimmed;
}

app.all("/api/health", (req, res) => {
  void sendFetchResponse(handleHealth, req, res);
});

app.all("/api/reports", (req, res) => {
  void sendFetchResponse(handleReports, req, res);
});

app.all("/api/events", (req, res) => {
  void sendFetchResponse(handleEvents, req, res);
});

app.all("/api/user/hidden-content", (req, res) => {
  void sendFetchResponse(handleUserHiddenContent, req, res);
});

app.all("/api/user/hidden-content/unhide", (req, res) => {
  void sendFetchResponse(handleUserHiddenContentUnhide, req, res);
});

app.all("/api/admin/reports", (req, res) => {
  void sendFetchResponse(handleAdminReports, req, res);
});

app.all("/api/admin/me", (req, res) => {
  void sendFetchResponse(handleAdminMe, req, res);
});

app.all("/api/admin/admins", (req, res) => {
  void sendFetchResponse(handleAdminRoster, req, res);
});

app.all("/api/admin/moderation/summary", (req, res) => {
  void sendFetchResponse(handleAdminModerationSummary, req, res);
});

app.all("/api/admin/ops/summary", (req, res) => {
  void sendFetchResponse(handleAdminOpsSummary, req, res);
});

app.all("/api/admin/users", (req, res) => {
  void sendFetchResponse(handleAdminUsers, req, res);
});

const analyticsRouteSegments = [
  "overview",
  "timeseries",
  "top-content",
  "search",
  "events",
  "health",
  "dead-letters",
  "aggregate",
] as const;

for (const segment of analyticsRouteSegments) {
  app.all(`/api/admin/analytics/${segment}`, (req, res) => {
    void sendFetchResponse(
      (request) => dispatchAdminAnalyticsApi(request, `admin/analytics/${segment}`),
      req,
      res,
    );
  });
}

app.all("/api/admin/analytics/events/:eventId", (req, res) => {
  void sendFetchResponse(
    (request) => dispatchAdminAnalyticsApi(request, `admin/analytics/events/${req.params.eventId}`),
    req,
    res,
  );
});

const businessAnalyticsSegments = [
  "overview",
  "growth",
  "funnel",
  "content",
  "search",
  "creators",
  "locations",
  "investor-snapshot",
] as const;

for (const segment of businessAnalyticsSegments) {
  app.all(`/api/admin/analytics/business/${segment}`, (req, res) => {
    void sendFetchResponse(
      (request) => dispatchAdminAnalyticsApi(request, `admin/analytics/business/${segment}`),
      req,
      res,
    );
  });
}

const businessIntelligenceSegments = [
  "dashboard",
  "overview",
  "geography",
  "markets",
  "places",
  "routes",
  "categories",
  "timeseries",
  "funnel",
  "searches",
  "opportunities",
  "content-attribution",
] as const;

for (const segment of businessIntelligenceSegments) {
  app.all(`/api/admin/business/${segment}`, (req, res) => {
    void sendFetchResponse(
      (request) => dispatchBusinessIntelligenceApi(request, `admin/business/${segment}`),
      req,
      res,
    );
  });
}

app.all("/api/cron/analytics/aggregate", (req, res) => {
  void sendFetchResponse(handleCronAnalyticsAggregate, req, res);
});

app.all("/api/pioneers/landing", (req, res) => {
  void sendFetchResponse(handlePioneersLanding, req, res);
});

app.all("/api/admin/system/health", (req, res) => {
  void sendFetchResponse(handleAdminSystemHealth, req, res);
});

app.all("/api/admin/system/metrics", (req, res) => {
  void sendFetchResponse(handleAdminSystemMetrics, req, res);
});

app.all("/api/admin/system/bootstrap-board", (req, res) => {
  void sendFetchResponse(handleBootstrapBoardAdmins, req, res);
});

const openApiSurfaces = ["postgrest", "edge", "admin"] as const;

for (const surface of openApiSurfaces) {
  app.all(`/api/admin/openapi/${surface}`, (req, res) => {
    void sendFetchResponse(
      (request) => dispatchOpenApiDocs(request, `admin/openapi/${surface}`),
      req,
      res,
    );
  });
}

app.all("/api/metrics", (req, res) => {
  void sendFetchResponse(handleTokenMetrics, req, res);
});

app.all("/api/admin/reports/:id", (req, res) => {
  void sendFetchResponse((request) => handleAdminReportById(request, req.params.id), req, res);
});

app.all("/api/admin/moderation/action", (req, res) => {
  void sendFetchResponse(handleAdminModerationAction, req, res);
});

app.all("/api/admin/broadcast", (req, res) => {
  void sendFetchResponse(adminBroadcast, req, res);
});

app.all("/api/feedback/submit", (req, res) => {
  void sendFetchResponse(feedbackSubmit, req, res);
});

/** Public waitlist signup (email only) — used by the Firebase front-end */
app.post("/api/waitlist/signup", async (req, res) => {
  try {
    const email = normalizeEmail(req.body?.email);
    const phone = typeof req.body?.phone === "string" ? req.body.phone.trim() : "";

    if (!email && !phone) {
      return res.status(400).json({ error: "A phone number or email is required." });
    }

    // Phone-only signup: Firestore (client) is the source of truth in dev; nothing to email.
    if (!email) {
      return res.json({ ok: true, created: true, message: "You're on the list." });
    }

    const { created } = addToWaitlist(email);
    try {
      await sendWaitlistEmail(email);
    } catch (mailErr) {
      console.error("[mail]", mailErr);
      return res.status(503).json({
        error: "Could not send confirmation email. Try again later.",
      });
    }

    return res.json({
      ok: true,
      created,
      message: created
        ? "You're on the list."
        : "You're already on the list.",
    });
  } catch (err) {
    console.error(err);
    return res.status(500).json({ error: "Something went wrong." });
  }
});

app.post("/api/access", async (req, res) => {
  try {
    const email = normalizeEmail(req.body?.email);
    const password = typeof req.body?.password === "string" ? req.body.password : undefined;

    if (!email) {
      return res.status(400).json({ error: "Invalid email address." });
    }

    if (email === config.adminEmail) {
      if (!password) {
        return res.json({ access: "password_required" as const });
      }
      if (password !== config.adminPassword) {
        return res.status(401).json({ error: "Incorrect password." });
      }
      const token = jwt.sign({ role: "admin", email }, config.jwtSecret, { expiresIn: "7d" });
      return res.json({ access: "full" as const, token });
    }

    const { created } = addToWaitlist(email);
    try {
      await sendWaitlistEmail(email);
    } catch (mailErr) {
      console.error("[mail]", mailErr);
      return res.status(503).json({
        error: "Could not send confirmation email. Is Mailpit running on port 1025?",
      });
    }

    return res.json({
      access: "waitlist" as const,
      created,
      message: created
        ? "You're on the list. Check your inbox — we'll notify you when the app is 100% ready."
        : "You're already on the list. We'll notify you when the app is 100% ready.",
    });
  } catch (err) {
    console.error(err);
    return res.status(500).json({ error: "Something went wrong. Please try again." });
  }
});

/** Admin: view who signed up (Firebase ID token from /team or legacy JWT) */
app.get("/api/admin/waitlist", requireAdmin, async (req, res) => {
  if (isFirebaseAdminConfigured()) {
    try {
      const data = await listWaitlistMergedAsync();
      return res.json(data);
    } catch (err) {
      console.error(err);
      return res.status(500).json({ error: "Could not load Firestore waitlist." });
    }
  }

  const limit = Math.min(Number(req.query.limit) || 100, 500);
  const offset = Math.max(Number(req.query.offset) || 0, 0);
  const stats = getWaitlistStats();
  const rows = listWaitlist(limit, offset);
  return res.json({ stats, rows, source: "local" });
});

/**
 * Admin: email everyone on the waitlist that the app is ready.
 * ?dryRun=1 — only returns who would receive it.
 * By default only emails rows where launch_notified_at IS NULL.
 */
app.post("/api/admin/waitlist/notify-launch", requireAdmin, async (req, res) => {
  const dryRun = req.query.dryRun === "1" || req.body?.dryRun === true;

  try {
    if (isFirebaseAdminConfigured()) {
      const pending = await listPendingLaunchEmailsMergedAsync();

      if (dryRun) {
        return res.json({ dryRun: true, count: pending.length, emails: pending });
      }

      if (pending.length === 0) {
        const { stats } = await listWaitlistMergedAsync();
        return res.json({ sent: [], failed: [], message: "No pending recipients.", stats });
      }

      const result = await sendAppLaunchBulk(pending, markLaunchNotifiedFirestore);
      const { stats } = await listWaitlistMergedAsync();
      return res.json({ ...result, stats, message: `Sent ${result.sent.length} launch email(s).` });
    }

    const pending = listWaitlistPendingLaunch();

    if (dryRun) {
      return res.json({
        dryRun: true,
        count: pending.length,
        emails: pending.map((r) => r.email),
      });
    }

    if (pending.length === 0) {
      return res.json({ sent: [], failed: [], message: "No pending recipients." });
    }

    const result = await sendAppLaunchBulk(
      pending.map((r) => r.email),
      markLaunchNotified,
    );
    const message =
      result.failed.length > 0
        ? `Sent ${result.sent.length}. Failed ${result.failed.length}.`
        : `Sent ${result.sent.length} launch email(s) via Mailpit/local SMTP.`;

    return res.json({
      ...result,
      stats: getWaitlistStats(),
      message,
    });
  } catch (err) {
    console.error(err);
    return res.status(500).json({ error: "Bulk send failed." });
  }
});

app.get("/api/me", (req, res) => {
  const header = req.headers.authorization;
  const token = header?.startsWith("Bearer ") ? header.slice(7) : null;
  if (!token) return res.status(401).json({ error: "Unauthorized" });

  try {
    const payload = jwt.verify(token, config.jwtSecret) as { role?: string; email?: string };
    if (payload.role !== "admin") return res.status(401).json({ error: "Unauthorized" });
    return res.json({ role: "admin", email: payload.email });
  } catch {
    return res.status(401).json({ error: "Unauthorized" });
  }
});

app.listen(config.port, () => {
  const obs = observabilityConfigStatus();
  const metricsConfigured = Boolean(String(process.env.METRICS_TOKEN || "").trim());
  logger.info("API server listening", {
    port: config.port,
    environment: process.env.APP_ENV || process.env.NODE_ENV || "development",
    metrics_token_configured: metricsConfigured,
    loki_ready: obs.loki_ready,
    grafana_logs_enabled: obs.grafana_logs_enabled,
  });
  console.log(`API http://localhost:${config.port}`);
  console.log(`Admin: ${config.adminEmail}`);
  console.log(`SMTP ${config.smtp.host}:${config.smtp.port} (Mailpit)`);
  console.log(
    `Observability: metrics=${metricsConfigured ? "on" : "off"} loki=${obs.loki_ready ? "ready" : "off"} → Grafana http://localhost:3002`,
  );
});